chore(actionpack): wave 3 — AbstractController split into abstractcontroller/ top-level dir
Extract AbstractController, ActionNotFound, and callback types from
actioncontroller/abstract-controller.ts into the Rails-mirrored
abstractcontroller/ top-level directory:

  abstractcontroller/base.ts      ← AbstractController class
  abstractcontroller/callbacks.ts ← ActionCallback, AroundCallback,
                                     CallbackOptions, CallbackEntry types
  abstractcontroller/error.ts     ← ActionNotFound

Keep actioncontroller/abstract-controller.ts as a backward-compat
re-export stub. Add AbstractController namespace export to the
package index.

Teach test:compare and generate-stubs about the new package dir;
add testPath for abstractcontroller in vendor/sources.ts pointing at
actionpack/test/abstract (3 Rails test files, 52 tests now tracked).

api:compare: abstractcontroller 4/82 methods, 2/11 files matched.
Overall delta: +1 method vs baseline.

Note: 639 LOC — exceeds 300-LOC ceiling due to single mechanical
file split (302-line source → 3 new files + stub). No logic changed.
