+++ packages/actionpack/src/abstractcontroller/index.ts
@@ -1,1 +1,8 @@
-export {};
+export {
+  AbstractController,
+  ActionNotFound,
+  type ActionCallback,
+  type AroundCallback,
+  type CallbackOptions,
+} from "./base.js";
+export type { CallbackEntry } from "./callbacks.js";
