+++ packages/actionpack/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./actiondispatch/index.js";
 export * as ActionController from "./actioncontroller/index.js";
+export * as AbstractController from "./abstractcontroller/index.js";
 export * as ActionView from "@blazetrails/actionview";
