+++ scripts/test-compare/generate-stubs.ts
@@ -26,6 +26,7 @@ const PKG_DIRS: Record<string, string> = {
   rack: "packages/rack/src/",
   actiondispatch: "packages/actionpack/src/actiondispatch/",
   actioncontroller: "packages/actionpack/src/actioncontroller/",
+  abstractcontroller: "packages/actionpack/src/abstractcontroller/",
   actionview: "packages/actionview/src/",
 };
